add module role and user, update all module

+++ src/photo_topic/photo_topic.service.ts
@@ -6,6 +6,8 @@ import { PhotoTopic } from './entities/photo_topic.entity';
 import { Repository } from 'typeorm';
 import { join } from 'path';
 import { unlink } from 'fs/promises';
+import * as fs from 'fs';
+import * as crypto from 'crypto';
 
 @Injectable()
 export class PhotoTopicService {
@@ -14,7 +16,18 @@ export class PhotoTopicService {
     private photoTopicRepository: Repository<PhotoTopic>,
   ) {}
 
-  async updatePhoto(id: number, files: Express.Multer.File) {
+   // Fungsi untuk Menghitung hash SHA-256 dari isi file
+   private calculateFileHash(filePath: string): string {
+    if (!fs.existsSync(filePath)) return '';
+
+    const fileBuffer = fs.readFileSync(filePath);
+    const hashSum = crypto.createHash('sha256');
+    hashSum.update(fileBuffer);
+
+    return hashSum.digest('hex');
+  }
+
+  async updatePhoto(id: string, files: Express.Multer.File) {
     // Cari foto berdasarkan id yang diberikan
     const photos = await this.photoTopicRepository.findOne({
       where: { id },
@@ -24,20 +37,35 @@ export class PhotoTopicService {
       throw new NotFoundException('Photo tidak ditemukan');
     }
 
-    const Path = join(process.cwd(), photos.file_path);
-    try {
-      await unlink(Path);
-    } catch (err) {
-      console.error('Error hapus data file foto: ', err);
+    if (files) {
+      const fileHash = this.calculateFileHash(files.path);
+      const existingHash = this.calculateFileHash(photos.file_path);
+
+      if (existingHash !== fileHash) {
+        // Hapus file cover lama di sistem
+        const Path = join(process.cwd(), photos.file_path);
+        try {
+          await unlink(Path); // Hapus file cover lama dari sistem
+        } catch (err) {
+          console.error('Error hapus data file foto: ', err);
+        }
+
+        // Ubah path cover dengan path yang baru
+        photos.file_path = files.path;
+
+        // Simpan file path foto yang baru
+        await this.photoTopicRepository.save(photos);
+        
+        return {
+          message: 'Foto berhasil diperbarui',
+          filePhoto: photos,
+        };
+      } else {
+        fs.unlinkSync(files.path);
+        return {
+          message: 'foto tidak ada yang baru',
+        };
+      }
     }
-
-    // Update path file di database
-    photos.file_path = files[0].path;
-    await this.photoTopicRepository.save(photos);
-
-    return {
-      message: 'Foto berhasil diperbarui',
-      filePhoto: photos,
-    };
   }
 }
